Feat/screen capture event listener for mac/win (#1015)
* Add screen sharing event listener.

* chore: force reload media list when re-getSources.

* chore: add stopRefershSources.

* chore: update.

* Update flutter_webrtc.podspec

* chore: update.

* fix typo.

* chore: Improve interface and events.

* chore: Improve and remove redundant code.

* chore: Remove code that is no longer used.

* feat: Add fps support for desktop capture.

* update.

* chore: Remove Timer from native layer.

* chore: print DesktopCapturer status.

* chore: Fix the thumbnail cannot pass value when getSources.

* using videoSourceForScreenCast create videoSource for screen sharing.

* chore: Compatible for flutter web.

+++ common/cpp/src/flutter_video_renderer.cc
@@ -43,7 +43,7 @@ const FlutterDesktopPixelBuffer* FlutterVideoRenderer::CopyPixelBuffer(
   if (pixel_buffer_.get() && frame_.get()) {
     if (pixel_buffer_->width != frame_->width() ||
         pixel_buffer_->height != frame_->height()) {
-      size_t buffer_size = (frame_->width() * frame_->height()) * (32 >> 3);
+      size_t buffer_size = (size_t(frame_->width()) * size_t(frame_->height())) * (32 >> 3);
       rgb_buffer_.reset(new uint8_t[buffer_size]);
       pixel_buffer_->width = frame_->width();
       pixel_buffer_->height = frame_->height();
